refactor and add last acq fields
