+++ recOrder/plugin/napari_plugin_entry_point.py
@@ -1,5 +1,5 @@
 from napari_plugin_engine import napari_hook_implementation
-from recOrder.plugin.calibration.calibration_module import recOrder_Calibration
+from recOrder.plugin.plugin_widget import recOrder_Widget
 
 """
 each of these GUI files is generated by qtdesigner.
@@ -10,19 +10,8 @@ pyuic5 -x <.ui input file> -o <.py output file>
 """
 
 
-# class RecOrderCalibration(QWidget):
-#     def __init__(self, napari_viewer):
-#         super().__init__()
-#         self.viewer = napari_viewer
-#
-#         self.ui = recOrder_calibration_v4.Ui_Form()
-#         self.ui.setupUi(self)
-#
-#         CalibrationModule(self.ui)
-
-
 @napari_hook_implementation
 def napari_experimental_provide_dock_widget():
     # you can return either a single widget, or a sequence of widgets
     # each widget is accessible as a new plugin that stacks in the side panel
-    return [recOrder_Calibration]
+    return [recOrder_Widget]
