+++ recOrder/plugin/qtdesigner/recOrder_calibration_v5.py
@@ -39,7 +39,7 @@ class Ui_Form(object):
         font.setWeight(50)
         self.tabWidget.setFont(font)
         self.tabWidget.setTabPosition(QtWidgets.QTabWidget.North)
-        self.tabWidget.setElideMode(QtCore.Qt.ElideRight)
+        self.tabWidget.setElideMode(QtCore.Qt.ElideLeft)
         self.tabWidget.setUsesScrollButtons(False)
         self.tabWidget.setTabsClosable(False)
         self.tabWidget.setMovable(False)
@@ -112,7 +112,7 @@ class Ui_Form(object):
         self.scrollArea_2.setWidgetResizable(True)
         self.scrollArea_2.setObjectName("scrollArea_2")
         self.scrollAreaWidgetContents_2 = QtWidgets.QWidget()
-        self.scrollAreaWidgetContents_2.setGeometry(QtCore.QRect(0, 0, 410, 300))
+        self.scrollAreaWidgetContents_2.setGeometry(QtCore.QRect(0, 0, 442, 300))
         self.scrollAreaWidgetContents_2.setObjectName("scrollAreaWidgetContents_2")
         self.verticalLayout_5 = QtWidgets.QVBoxLayout(self.scrollAreaWidgetContents_2)
         self.verticalLayout_5.setObjectName("verticalLayout_5")
@@ -283,6 +283,14 @@ class Ui_Form(object):
         self.qbutton_calibrate.setAutoDefault(False)
         self.qbutton_calibrate.setObjectName("qbutton_calibrate")
         self.verticalLayout_2.addWidget(self.qbutton_calibrate)
+        self.qbutton_calc_extinction = QtWidgets.QPushButton(self.run_calib)
+        font = QtGui.QFont()
+        font.setBold(False)
+        font.setItalic(False)
+        font.setWeight(50)
+        self.qbutton_calc_extinction.setFont(font)
+        self.qbutton_calc_extinction.setObjectName("qbutton_calc_extinction")
+        self.verticalLayout_2.addWidget(self.qbutton_calc_extinction)
         self.label_progress = QtWidgets.QLabel(self.run_calib)
         font = QtGui.QFont()
         font.setBold(False)
@@ -312,14 +320,6 @@ class Ui_Form(object):
         self.le_extinction.setReadOnly(True)
         self.le_extinction.setObjectName("le_extinction")
         self.verticalLayout_2.addWidget(self.le_extinction)
-        self.qbutton_calc_extinction = QtWidgets.QPushButton(self.run_calib)
-        font = QtGui.QFont()
-        font.setBold(False)
-        font.setItalic(False)
-        font.setWeight(50)
-        self.qbutton_calc_extinction.setFont(font)
-        self.qbutton_calc_extinction.setObjectName("qbutton_calc_extinction")
-        self.verticalLayout_2.addWidget(self.qbutton_calc_extinction)
         self.labell_calib_assessment = QtWidgets.QLabel(self.run_calib)
         font = QtGui.QFont()
         font.setBold(False)
@@ -329,6 +329,7 @@ class Ui_Form(object):
         self.labell_calib_assessment.setObjectName("labell_calib_assessment")
         self.verticalLayout_2.addWidget(self.labell_calib_assessment)
         self.le_calib_assessment = QtWidgets.QLineEdit(self.run_calib)
+        self.le_calib_assessment.setReadOnly(True)
         self.le_calib_assessment.setObjectName("le_calib_assessment")
         self.verticalLayout_2.addWidget(self.le_calib_assessment)
         self.gridLayout.addWidget(self.run_calib, 4, 0, 1, 1)
@@ -709,9 +710,9 @@ class Ui_Form(object):
         self.cb_calib_scheme.setItemText(1, _translate("Form", "5-State (Ext, 0, 45, 90, 135)"))
         self.run_calib.setTitle(_translate("Form", "Run Calibration"))
         self.qbutton_calibrate.setText(_translate("Form", "Run Calibration"))
+        self.qbutton_calc_extinction.setText(_translate("Form", "Calculate Extinction"))
         self.label_progress.setText(_translate("Form", "Progress"))
         self.label_extinction.setText(_translate("Form", "Extinction"))
-        self.qbutton_calc_extinction.setText(_translate("Form", "Calculate Extinction"))
         self.labell_calib_assessment.setText(_translate("Form", "Calibration Assessment"))
         self.capture_background.setTitle(_translate("Form", "Capture Background"))
         self.label_bg_folder.setText(_translate("Form", "Background Folder Name"))
