+++ recOrder/plugin/qtdesigner/recOrder_calibration_v5.ui
@@ -66,7 +66,7 @@
       <number>0</number>
      </property>
      <property name="elideMode">
-      <enum>Qt::ElideRight</enum>
+      <enum>Qt::ElideLeft</enum>
      </property>
      <property name="usesScrollButtons">
       <bool>false</bool>
@@ -218,7 +218,7 @@
                    <rect>
                     <x>0</x>
                     <y>0</y>
-                    <width>410</width>
+                    <width>442</width>
                     <height>300</height>
                    </rect>
                   </property>
@@ -537,6 +537,20 @@
                </widget>
               </item>
               <item>
+               <widget class="QPushButton" name="qbutton_calc_extinction">
+                <property name="font">
+                 <font>
+                  <weight>50</weight>
+                  <italic>false</italic>
+                  <bold>false</bold>
+                 </font>
+                </property>
+                <property name="text">
+                 <string>Calculate Extinction</string>
+                </property>
+               </widget>
+              </item>
+              <item>
                <widget class="QLabel" name="label_progress">
                 <property name="font">
                  <font>
@@ -585,20 +599,6 @@
                </widget>
               </item>
               <item>
-               <widget class="QPushButton" name="qbutton_calc_extinction">
-                <property name="font">
-                 <font>
-                  <weight>50</weight>
-                  <italic>false</italic>
-                  <bold>false</bold>
-                 </font>
-                </property>
-                <property name="text">
-                 <string>Calculate Extinction</string>
-                </property>
-               </widget>
-              </item>
-              <item>
                <widget class="QLabel" name="labell_calib_assessment">
                 <property name="font">
                  <font>
@@ -613,7 +613,11 @@
                </widget>
               </item>
               <item>
-               <widget class="QLineEdit" name="le_calib_assessment"/>
+               <widget class="QLineEdit" name="le_calib_assessment">
+                <property name="readOnly">
+                 <bool>true</bool>
+                </property>
+               </widget>
               </item>
              </layout>
             </widget>
